requirements.txt
Django ~= 4.2.3
djangorestframework >= 3.15.2
celery >= 5.2.7, < 6
django-bootstrap-form >= 3.4, < 4
oauthlib >= 3.2.2, < 4
requests >= 2.28.2, < 3
git+https://github.com/apluslms/django-lti-login.git@3.3.0#egg=django-lti-login==3.3.0
git+https://github.com/apluslms/django-essentials.git@1.6.0#egg=django-essentials==1.6.0
git+https://github.com/apluslms/a-plus-client.git@v1.3.0#egg=a_plus_client
aplus-auth ~= 0.2.3
setuptools >= 78.1.1
django-debug-toolbar ~= 4.4.6
pytz ~= 2024.2
pylibmc ~= 1.6.3
pytest-playwright ~= 0.6.2
pytest-django ~= 4.9.0
pytest-env ~= 1.1.5
Pygments ~= 2.19.1
flower ~= 2.0.1
django-revproxy ~= 0.13.0
